refactor: Phase 4 - Move PLT resolution to arch/aarch64::plt
Move PLT (Procedure Linkage Table) resolution functions and tests from
call_graph.rs to arch/aarch64::plt submodule:

- Move build_plt_map() -> arch::plt::build_map()
- Move resolve_plt_stub() -> arch::plt::resolve_stub()
- Move 5 PLT-related tests to arch/aarch64::plt::tests
- Remove unused goblin::elf::Elf import from call_graph.rs

PLT resolution is ELF-specific and ARM64-specific (uses fixed 16-byte
entry size and 32-byte resolver stub). Grouping it with other ARM64
code improves organization.

All 405 tests passing.
